macos: configure omnibus for universal builds
Requires a recompile of LLVM compiled for universal
(i.e. both archs in every binary and library) and
results in roughly double the package size but with
the benefit that the embedded distribution can be
used on any mac of an archtecture and be used for
cross-compilation to any other Mac architecture.