Fix target triple on macOS
With the release of Xcode 15, a new linker was introduced and with it
came warnings from the linker. It appears that the new linker is
stricter in which target triples it considers valid. Specifically, it
looks like the minimum deployment target is required. E.g.,
`aarch64-apple-darwin23.3.0` is valid, while `aarch64-apple-darwin` is
not.

See also:

* https://developer.apple.com/documentation/xcode-release-notes/xcode-15-release-notes#Linking
* https://github.com/crystal-lang/crystal/issues/13846