Make the function work upon viewer request
CloudFront doesn't run viewer response functions when the origin
responds a 4xx or 5xx, so we have to hard-code which requests can pass
straight to the origin.

This function will need to get updated if we add extra files to the
`/api/` directory, such as a 2.x release, or other new files - I predict
this will be the source of a few bugs in the future.