Build architecture-agnostic docker images (#390)
Previously, we copied a specific tarball into the docker image. This doesn't work well for multiarch builds because we need different tarballs depending on the target architecture.
With this change, we copy the appropriate tarball from a directory which may contain tarballs for multiple architectures.
This changes the CLI interface for `docker/Makefile`: It no longer receives a `CRYSTAL_TARGZ` path, but a path to a directory with the tarball(s).
The refactoring also simplifies the Makefile a little bit (but more to come).